# openapi.yaml
openapi: 3.0.3
info:
  title: Defellix API
  description: >
    API Documentation for the Defellix Decentralized Freelancer Trust Platform.
    This API gateway routes requests to underlying microservices: Auth, User, Contract, and Blockchain.
  version: 1.0.0
servers:
  - url: https://api.defellix.com
    description: Production API via Cloudflare Tunnels
  - url: http://localhost:8080
    description: Local development via Nginx API Gateway
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token in the format **Bearer &lt;token&gt;**
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, full_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        full_name:
          type: string
          minLength: 2
          maxLength: 100
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        user_id:
          type: integer
        email:
          type: string

    # --- USER MODULE SCHEMAS ---
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/auth/oauth/google:
    get:
      tags: [Auth]
      summary: Login with Google (OAuth 2.0)
      description: Redirects the user to Google's consent screen. Set the `role` parameter to determine the account type for new registrations.
      parameters:
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum: [freelancer, client]
            default: freelancer
      responses:
        '307':
          description: Temporary redirect to Google Identity Provider
  /api/v1/auth/oauth/linkedin:
    get:
      tags: [Auth]
      summary: Login with LinkedIn (OAuth 2.0)
      description: Redirects the user to LinkedIn's consent screen. Set the `role` parameter to determine the account type for new registrations.
      parameters:
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum: [freelancer, client]
            default: freelancer
      responses:
        '307':
          description: Temporary redirect to LinkedIn Identity Provider
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User details returned

  # ==========================================
  # USER SERVICE
  # ==========================================
...
  /api/v1/contracts/{id}/certificate:
    get:
      tags: [Contract]
      summary: GET Section 65B Certificate
      description: Auto-generates a court-admissible HTML/PDF document citing blockchain proof and contract terms. Only available for `signed` or `completed` contracts.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: HTML document returned for printing to PDF.

  # ==========================================
  # WEb3 WALLET & BLOCKCHAIN SERVICES
  # ==========================================
...